% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_panel_modal.R
\name{controlPanelModal}
\alias{controlPanelModal}
\title{Generate Control Panel Modal}
\usage{
controlPanelModal(values, title)
}
\arguments{
\item{values}{list of values to provide by default}

\item{title}{Header title for the modal}
}
\value{
Return Dialog modal for control panel.
}
\description{
Generate Control Panel Modal
}
