% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{createExperiment}
\alias{createExperiment}
\title{Create experiment class object}
\usage{
createExperiment(folder_path)
}
\arguments{
\item{folder_path}{File input}
}
\value{
Experiment Class
}
\description{
Wrapper for R6 method
}
\examples{
\dontrun{
e <- createExperiment(folder_path = './')
}
}
