% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createFragUploadFolder}
\alias{createFragUploadFolder}
\title{Create the folder one would need to upload to fragalysis}
\usage{
createFragUploadFolder(data, copymaps = FALSE)
}
\arguments{
\item{data}{an experiment object}

\item{copymaps}{bool, if you want to copy the maps}
}
\value{
Creates some folders on the file system and return the filepath of the zipped file!
}
\description{
Create the folder one would need to upload to fragalysis
}
