% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Experiment}
\alias{Experiment}
\title{Experiment Class}
\description{
Experiment Class

Experiment Class
}
\details{
Describes a collection of ligands from fragalysis-api output.
}
\examples{
\dontrun{
e <- Experiment$new(folder_path = './')
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{target}}{Name of protein system}

\item{\code{ligands}}{List of Ligands}

\item{\code{filepath}}{input directory}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{get_ligands}}{Vector of ligand names}

\item{\code{get_apo_files}}{Vector of apo file names}

\item{\code{get_mol_files}}{Vector of .mol files}

\item{\code{get_2fofc_maps}}{Vector of 2fofc maps}

\item{\code{get_fofc_maps}}{Vector of fofc maps}

\item{\code{get_event_maps}}{Vector of event maps}

\item{\code{get_smiles}}{Vector of SMILES strings}

\item{\code{get_reviews}}{Table of Reviews}

\item{\code{get_metadata}}{Table of meta data}

\item{\code{get_status}}{Get the review status of ligands}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Experiment$new()}}
\item \href{#method-clone}{\code{Experiment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Experiment object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$new(folder_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{folder_path}}{Directory path}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
experiment class!
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
