% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_ui.R
\name{flex_ui}
\alias{flex_ui}
\title{Flexible UI sidebar panel for XCR}
\usage{
flex_ui(which, session_data)
}
\arguments{
\item{which}{The name of the tab}

\item{session_data}{Some reactive values}
}
\value{
Corresponding UI elements for sidebar
}
\description{
Flexible UI sidebar panel for XCR
}
