% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_residues}
\alias{get_residues}
\title{Get residue list from pdb_file}
\usage{
get_residues(pdb_file)
}
\arguments{
\item{pdb_file}{PDB file}
}
\value{
Returns list of resiudes
}
\description{
Get residue list from pdb_file
}
