% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getCurrentParams}
\alias{getCurrentParams}
\title{Get current values from shiny input object}
\usage{
getCurrentParams(input)
}
\arguments{
\item{input}{a Shiny input object usually within server}
}
\value{
a List containing values from specified fields
}
\description{
Get current values from shiny input object
}
