% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getExt}
\alias{getExt}
\title{Get the file extension of a string (if any).}
\usage{
getExt(x)
}
\arguments{
\item{x}{A string}
}
\value{
A string corresponding the file extension
}
\description{
Get the file extension of a string (if any).
}
\examples{
getExt('afile.txt')
}
