% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{getLigandProp}
\alias{getLigandProp}
\title{Get a specific property from a list of ligands}
\usage{
getLigandProp(value, ligands, what)
}
\arguments{
\item{value}{here for decoration, not needed}

\item{ligands}{the list of ligands}

\item{what}{what property to get}
}
\value{
gets some stuff from ligands i guess
}
\description{
Get a specific property from a list of ligands
}
\examples{
\dontrun{
ligands <- list(
 Ligand$new(name = 'Prot-x1234', base = './'),
 Ligand$new(name = 'Prot-x1233', base = './')
)
getLigandProp(ligands=ligands, what = 'name')
}
}
