% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Ligand}
\alias{Ligand}
\title{Ligand Class}
\description{
Ligand Class

Ligand Class
}
\details{
Describes a single Ligand from fragalysis-api
}
\examples{
\dontrun{
l <- Ligand$new(name = 'Prot-x1234', base = './')
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Name of the ligand}

\item{\code{crystal}}{Name of the crystal the ligand belongs to}

\item{\code{aligned_loc}}{path of the aligned folder}

\item{\code{metadata_filepath}}{path to the metadata file}

\item{\code{crys_loc}}{path of the crystallographic folder}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{apo_pdb_file}}{file path where apo.pdb should exist}

\item{\code{mol_file}}{file path where .mol should exist}

\item{\code{twofofc_file}}{file path where _2fofc.map should exist}

\item{\code{fofc_file}}{file path where _fofc.map should exist}

\item{\code{smiles_file}}{file path where _smiles.txt should exist}

\item{\code{smiles_string}}{The contents of _smiles.txt}

\item{\code{event_map_paths}}{file paths where event maps should reside}

\item{\code{first_event_map}}{returns the first eventmap}

\item{\code{metadata}}{the contents of the metadata}

\item{\code{review_file}}{file path where _review.txt should exist}

\item{\code{review}}{Contents of the review file}

\item{\code{crys_pdb_file}}{file path where crystallographic pdb should exist}

\item{\code{crys_event_maps}}{file path where crystallographic event exist}

\item{\code{crys_first_event_map}}{file path of 1st event map}

\item{\code{crys_2fofc_map}}{filepath of 2fofc map}

\item{\code{crys_fofc_map}}{filepath of fofc map}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Ligand$new()}}
\item \href{#method-clone}{\code{Ligand$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new ligand
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ligand$new(name, base)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of Ligand}

\item{\code{base}}{Directory path containing ligandy goodness}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ligand class!
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ligand$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
