% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loadDefaultParams}
\alias{loadDefaultParams}
\title{Get XCR specific defaults}
\usage{
loadDefaultParams()
}
\value{
A list of default values for fogging, clipping, boxsize, 
 clipDist, backgroundColor, cameraType and mousePreset
}
\description{
Get XCR specific defaults
}
