% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{n_anno}
\alias{n_anno}
\title{Number of ligands annotated}
\usage{
n_anno(exp)
}
\arguments{
\item{exp}{Experiment Class}
}
\value{
Integer, corresponding to number of ligands annotated
}
\description{
The number of annotated ligands in an experiment
}
\examples{
\dontrun{
e <- createExperiment(folder_path = './')
n_anno(e)
}
}
