% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{n_ligs}
\alias{n_ligs}
\title{Number of ligands}
\usage{
n_ligs(exp)
}
\arguments{
\item{exp}{Experiment Class}
}
\value{
Integer, corresponding to number of ligands
}
\description{
The number of ligands in an experiment
}
\examples{
\dontrun{
e <- createExperiment(folder_path = './')
n_ligs(e)
}
}
