% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{n_revi}
\alias{n_revi}
\title{Number of ligands reviewed}
\usage{
n_revi(exp)
}
\arguments{
\item{exp}{Experiment Class}
}
\value{
Integer, corresponding to number of ligands that are reviewed
}
\description{
The number of reviewed ligands in an experiment
}
\examples{
\dontrun{
e <- createExperiment(folder_path = './')
n_revi(e)
}
}
