% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normalise_path}
\alias{normalise_path}
\title{Normalise shinyFiles path}
\usage{
normalise_path(x, root = getwd())
}
\arguments{
\item{x}{The shinyFiles output}

\item{root}{The start of the file path}
}
\value{
Character String of the normalised path
}
\description{
Converts a shinyFiles path to an absolute path which should be valid given
a particular root
}
