% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pathOrNA}
\alias{pathOrNA}
\title{Return a valid filepath or a NA Value}
\usage{
pathOrNA(path)
}
\arguments{
\item{path}{A string specifying a filepath to test}
}
\value{
a string or NA value
}
\description{
Return a valid filepath or a NA Value
}
\examples{
pathOrNA('NotAPath/NotAValidPath')
pathOrNA('./') # Valid
}
