% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remapData}
\alias{remapData}
\title{Create reactive environment - from reactive values}
\usage{
remapData(session_data)
}
\arguments{
\item{session_data}{Some reactive values}
}
\value{
Return review table in reactive object
}
\description{
Create reactive environment - from reactive values
}
