% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remapFragviewData}
\alias{remapFragviewData}
\title{Create another reactive environment - from reactive values}
\usage{
remapFragviewData(session_data)
}
\arguments{
\item{session_data}{Some reactive values}
}
\value{
Return fragview table in reactive object
}
\description{
Create another reactive environment - from reactive values
}
