% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{removeNamedComponent}
\alias{removeNamedComponent}
\title{Remove component from NGL stage}
\usage{
removeNamedComponent(session, objectname)
}
\arguments{
\item{session}{A shiny server session object}

\item{objectname}{window.object to get rid of}
}
\value{
Returns Nothing
}
\description{
Remove component from NGL stage
}
