% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{removeSpaces}
\alias{removeSpaces}
\title{Escape spaces in a string}
\usage{
removeSpaces(x)
}
\arguments{
\item{x}{The string to remove spaces from}
}
\value{
x but with `\\ ` instead of ` `
}
\description{
This function replaces all single white space characters with escaped 
versions (e.g. `\\ `)
}
\examples{
removeSpaces('Remove Space')
}
