% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{resetForm}
\alias{resetForm}
\title{Reset the review form}
\usage{
resetForm(session, session_data)
}
\arguments{
\item{session}{Shiny Session object}

\item{session_data}{Some reactive values}
}
\value{
Returns nothing
}
\description{
Reset the review form
}
