% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_ui.R
\name{review_ui}
\alias{review_ui}
\title{Generate Review UI components}
\usage{
review_ui(session_data)
}
\arguments{
\item{session_data}{Some reactive values}
}
\value{
Corresponding UI elements for Review
}
\description{
Generate Review UI components
}
