% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rsplit}
\alias{rsplit}
\title{Split a starting from the right}
\usage{
rsplit(string, split_by, n = 1)
}
\arguments{
\item{string}{The string to split}

\item{split_by}{The character to split string by}

\item{n}{The number of split you want to perform}
}
\value{
The string split from the right
}
\description{
This is an R equivalent to the python str.rsplit method which will return a
vector of length n+1 of a string split by a given character starting from 
the right hand side of the string
}
\examples{
rsplit(string = 'my_string_abc', split_by = '_', n = 1)
rsplit(string = 'my_string_abc', split_by = '_', n = 2)
}
