% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{saveReview}
\alias{saveReview}
\title{Save review of ligand to filesystem}
\usage{
saveReview(x, z, ligand)
}
\arguments{
\item{x}{Vector of review info...}

\item{z}{Atom crique}

\item{ligand}{ligand class object}
}
\value{
Returns nothing but will edit a couple of files
}
\description{
Save review of ligand to filesystem
}
