% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tcl}
\alias{tcl}
\title{Convert a logical value into lower case characters for java-script}
\usage{
tcl(x)
}
\arguments{
\item{x}{a bool value}
}
\value{
Either 'true' or 'false'
}
\description{
Converts a TRUE/FALSE value into 'true' or 'false', for some reason I need
this and others might too.
}
\examples{
tcl(TRUE)
tcl(FALSE)
}
