% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{updateDensityBoxSize}
\alias{updateDensityBoxSize}
\title{Update the boxsize level of a map}
\usage{
updateDensityBoxSize(session, name, boxsize)
}
\arguments{
\item{session}{A shiny server session object}

\item{name}{name of the window to update}

\item{boxsize}{numeric, value to update boxsize to}
}
\value{
Returns Nothing
}
\description{
Update the boxsize level of a map
}
