% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{updateDensityISO}
\alias{updateDensityISO}
\title{Update the ISO level of a map}
\usage{
updateDensityISO(session, name, isolevel)
}
\arguments{
\item{session}{A shiny server session object}

\item{name}{name of the window to update}

\item{isolevel}{numeric, value to update iso level to}
}
\value{
Returns Nothing
}
\description{
Update the ISO level of a map
}
