% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{updateMainTable}
\alias{updateMainTable}
\title{Create a DT Render from a reactive data object}
\usage{
updateMainTable(x, pl = 25, format = TRUE)
}
\arguments{
\item{x}{data.frame inside a reactive environment}

\item{pl}{default row-numbers/pagelength of table}

\item{format}{Boolean, indicate whether special formatting rules should be applied}
}
\value{
An render instance to be supplied to shiny output
}
\description{
Internal function which will update a given table using some optional
formatting - to be suppled to a output$table
}
