% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{updateParam}
\alias{updateParam}
\title{Update NGL stage using updateaparam}
\usage{
updateParam(session, which, what)
}
\arguments{
\item{session}{A shiny server session object}

\item{which}{which parameter to change}

\item{what}{what value to change to}
}
\value{
Returns Nothing
}
\description{
Update NGL stage using updateaparam
}
