% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{updateVisability}
\alias{updateVisability}
\title{Toggle the visability of a map}
\usage{
updateVisability(session, name, bool)
}
\arguments{
\item{session}{A shiny server session object}

\item{name}{name of the window to update}

\item{bool}{bool, Which state the update to}
}
\value{
Returns Nothing
}
\description{
Toggle the visability of a map
}
