% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{uploadApoPDB}
\alias{uploadApoPDB}
\title{Upload an apo PDB to NGL Stage}
\usage{
uploadApoPDB(session, filepath, repr)
}
\arguments{
\item{session}{A shiny server session object}

\item{filepath}{The file to upload}

\item{repr}{Representation for th structure (e.g. line or cartoon)}
}
\value{
Returns Nothing
}
\description{
Upload an apo PDB to NGL Stage
}
