% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{uploadMF2}
\alias{uploadMF2}
\title{Upload a mol file and do something different to NGL Stage}
\usage{
uploadMF2(session, filepath)
}
\arguments{
\item{session}{A shiny server session object}

\item{filepath}{The file to upload}
}
\value{
Returns Nothing
}
\description{
Upload a mol file and do something different to NGL Stage
}
