% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{uploadMolAndFocus}
\alias{uploadMolAndFocus}
\title{Upload a mol file to NGL Stage (and focus!)}
\usage{
uploadMolAndFocus(session, filepath, ext)
}
\arguments{
\item{session}{A shiny server session object}

\item{filepath}{The file to upload}

\item{ext}{The file extension}
}
\value{
Returns Nothing
}
\description{
Upload a mol file to NGL Stage (and focus!)
}
