% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{uploadMolAndFocus2}
\alias{uploadMolAndFocus2}
\title{Upload a mol file and do something different to NGL Stage}
\usage{
uploadMolAndFocus2(session, filepath)
}
\arguments{
\item{session}{Shiny Session object}

\item{filepath}{The file to upload}
}
\value{
Returns Nothing
}
\description{
Upload a mol file and do something different to NGL Stage
}
