% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{uploadPDB}
\alias{uploadPDB}
\title{Upload a PDB to NGL Stage}
\usage{
uploadPDB(session, filepath, input)
}
\arguments{
\item{session}{A shiny server session object}

\item{filepath}{The file to upload}

\item{input}{A shiny server input object}
}
\value{
Returns Nothing
}
\description{
Upload a PDB to NGL Stage
}
