% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{uploadUnfocussedMol}
\alias{uploadUnfocussedMol}
\title{Upload a mol file and not focus to NGL Stage}
\usage{
uploadUnfocussedMol(session, filepath)
}
\arguments{
\item{session}{A shiny server session object}

\item{filepath}{The file to upload}
}
\value{
Returns Nothing
}
\description{
Upload a mol file and not focus to NGL Stage
}
