% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{uploadVolumeDensity}
\alias{uploadVolumeDensity}
\title{Upload a mol file and do something different to NGL Stage}
\usage{
uploadVolumeDensity(
  session,
  filepath,
  color,
  negateiso = FALSE,
  boxsize,
  isolevel,
  visable,
  windowname
)
}
\arguments{
\item{session}{A shiny server session object}

\item{filepath}{The file to upload}

\item{color}{string, colour of the map}

\item{negateiso}{bool, indicate whether or not iso should be negated
for fofc map files particularly}

\item{boxsize}{numeric, how large the map should render - 0 for max size}

\item{isolevel}{numeric, sigma contour level}

\item{visable}{bool, whether it should render or not (really?)}

\item{windowname}{string, name of the window object}
}
\value{
Returns Nothing
}
\description{
Upload a mol file and do something different to NGL Stage
}
